requirements.txt
# ─── Core GCP ────────────────────────────────────────────────
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2
google-cloud-dataproc==5.8.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# ─── Data Processing ─────────────────────────────────────────
pandas==2.2.1
pyarrow==15.0.0
numpy==1.26.4

# ─── API / Networking ────────────────────────────────────────
requests==2.31.0
urllib3==2.2.1
tenacity==8.2.3          # Retry logic for Tiingo API

# ─── Config / Utilities ──────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1

# ─── Dev / Test ──────────────────────────────────────────────
pytest==8.1.1
pytest-cov==5.0.0
flake8==7.0.0
black==24.3.0
